Factor public types into types.ts
