+++ src/index.ts
@@ -1,7 +1,7 @@
 import createMarker from './createMarker';
 export { default as mark } from './mark';
 
-export type { Parser } from './createMarker';
+export type { Parser, TagFunction } from './types';
 
 
 export default createMarker;
